Migrated to Spring Data JDBC
The intention of this change was to trial Spring Data JDBC, a DDD inspired repository framework in Spring. I found this to be so pleasant to use that I carried on to the finish.

It auto-generates CRUD style code for use and forces us towards the DDD repository pattern. We have the escape hatch of JDBI available to act as a data-mapper if need be, but as you can see in the `AccountRepository` the framework includes a `@Query` annotation that lets us override the behaviour of the framework when desirable.