Fixing the hardcoded localhost variable
I've also smuggled in some refactoring here. There's a further refactor needed to move more of the logic into the email viewmodels so that we don't have to attempt to test based off of teh email views. Alternatively, we can add regression tests to the email views where we save the generated HTML email in test resources and require them to be "re-recorded" on a change.