Admin users can make modifications to user settings
This means that an admin can change a user's name, email or handle without impersonating them.

While this feature (and the structure) seem like overkill, the follow-up to this change is to give admins the ability to change a user's account status (e.g. to freeze or block an account).